refactor(dispatcher): introduce LogBuilder and split module into sub-files
P0 prep2 – LogBuilder fluent API:
- Add CallCtx / CacheWriteCtx / RequestExtras to consolidate 20-30
  parameter handler signatures down to 6-10 parameters
- Add LogBuilder with from_ctx / from_dispatch constructors and
  chainable methods; replace all 23 emit_log positional-arg call
  sites across dispatch_pipeline, dispatch, and all three handlers

P0 – physical module split (mv + visibility only, zero logic change):
- auth.rs        API-key auth + per-route quota enforcement
- non_stream.rs  handle_non_stream + handle_non_stream_via_upstream_stream
- stream.rs      handle_stream (passthrough + IR round-trip)
- mod.rs retains entry points, context/log types, response helpers,
  and semantic-embedding computation (~1238 lines, down from 2072)

Zero warnings, zero behaviour change.
